Consolidate two internal files
